--- a/benches/metrics_no_tracing_integration.rs
+++ b/benches/metrics_no_tracing_integration.rs
@@ -3,7 +3,7 @@ use criterion::{criterion_group, criterion_main, Criterion};
 mod metrics_bench_util;
 
 fn benchmark(c: &mut Criterion) {
-    metrics_bench_util::benchmark(c, metrics_bench_util::Mode::MetricsNoTracingLayer)
+    metrics_bench_util::benchmark(c, metrics_bench_util::Mode::MetricsNoTracingIntegration)
 }
 
 criterion_group!(benches, benchmark);