--- a/benches/metrics_on.rs
+++ b/benches/metrics_on.rs
@@ -1,0 +1,10 @@
+use criterion::{criterion_group, criterion_main, Criterion};
+
+mod metrics_bench_util;
+
+fn benchmark(c: &mut Criterion) {
+    metrics_bench_util::benchmark(c, true)
+}
+
+criterion_group!(benches, benchmark);
+criterion_main!(benches);