This PR adds benchmarks to the metrics system.

It also
- introduces some (hidden and undocumented) escape hatches to disable internal metrics subsystem,
- vendors the metrics registry implementation (to allow more flexibility in integrating with our needs),
- optimizes metrics snapshot procedure,
- disables `autobenches` cargo option (with our current configuration it only confuses people),
- adds the ability to reset the metrics registry via a controller.

Closes #5093.

---

Findings from the benches show that the topology of socket source to socket sink performs slower with internal metrics than without them by the order of roughly up to 10-15%, but ultimately the performance change *within the noize tolerance*.

```shell
$ critcmp | grep topology
metrics_no_tracing_layer/topology/tcp_socket/01_writer             1.00     22.9±4.31ms    41.7 MB/sec    1.00     22.9±4.31ms    41.7 MB/sec
metrics_no_tracing_layer/topology/tcp_socket/02_writers            1.00    32.0±11.30ms    59.5 MB/sec    1.00    32.0±11.30ms    59.5 MB/sec
metrics_no_tracing_layer/topology/tcp_socket/04_writers            1.00    68.5±36.47ms    55.7 MB/sec    1.00    68.5±36.47ms    55.7 MB/sec
metrics_no_tracing_layer/topology/tcp_socket/08_writers            1.00    112.0±3.36ms    68.1 MB/sec    1.00    112.0±3.36ms    68.1 MB/sec
metrics_no_tracing_layer/topology/tcp_socket/16_writers            1.00    239.3±4.77ms    63.8 MB/sec    1.00    239.3±4.77ms    63.8 MB/sec
metrics_off/topology/tcp_socket/01_writer                          1.00     20.8±5.73ms    45.9 MB/sec    1.00     20.8±5.73ms    45.9 MB/sec
metrics_off/topology/tcp_socket/02_writers                         1.00    31.1±12.19ms    61.3 MB/sec    1.00    31.1±12.19ms    61.3 MB/sec
metrics_off/topology/tcp_socket/04_writers                         1.00     53.5±8.24ms    71.3 MB/sec    1.00     53.5±8.24ms    71.3 MB/sec
metrics_off/topology/tcp_socket/08_writers                         1.00    114.4±6.01ms    66.7 MB/sec    1.00    114.4±6.01ms    66.7 MB/sec
metrics_off/topology/tcp_socket/16_writers                         1.00    235.5±3.11ms    64.8 MB/sec    1.00    235.5±3.11ms    64.8 MB/sec
metrics_on/topology/tcp_socket/01_writer                           1.00     23.7±5.59ms    40.2 MB/sec    1.00     23.7±5.59ms    40.2 MB/sec
metrics_on/topology/tcp_socket/02_writers                          1.00    32.7±11.48ms    58.3 MB/sec    1.00    32.7±11.48ms    58.3 MB/sec
metrics_on/topology/tcp_socket/04_writers                          1.00    58.8±22.47ms    64.8 MB/sec    1.00    58.8±22.47ms    64.8 MB/sec
metrics_on/topology/tcp_socket/08_writers                          1.00    112.8±5.26ms    67.6 MB/sec    1.00    112.8±5.26ms    67.6 MB/sec
metrics_on/topology/tcp_socket/16_writers                          1.00    240.4±4.67ms    63.5 MB/sec    1.00    240.4±4.67ms    63.5 MB/sec
```