@MOZGIII 2020-12-23
I think it's already running at CI with the rest of them - I've wired it up through the Makefile, and it should already be showing up in Github Actions.